requirements.txt
configparser~=7.2.0
openai~=1.107.2
python-dotenv~=1.1.1
requests~=2.32.4
youtube-transcript-api~=1.2.2